# agent/api.yaml
openapi: "3.0.0"
info:
  version: "0.1.0"
  title: "Squill Agent REST API"
  description: The REST API used by the Squill Client to interact with the Agent.
paths:
  /agent:
    get:
      summary: Agent info.
      description: Get the information about the running agent.
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "401":
          description: Unauthorized
...
  /users/{username}/user:
    get:
      parameters:
        - name: username
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Username"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
        "401":
          description: Unauthorized
components:
  schemas:
    Agent:
      description: Description of the agent.
      type: object
      required:
        - version
      properties:
        version:
          type: string
...
    User:
      description: A user that can logon to the system.
      type: object
      required:
        - username
      properties:
        username:
          type: string
...
    Username:
      description: A username
      type: string
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key

    BearerAuth:
      type: http
      scheme: bearer
security:
  - ApiKeyAuth: []
  - BearerAuth: []